refactor: split up get.rs
