+++ grovedb/src/operations/get/worst_case.rs
@@ -1,0 +1,61 @@
+use costs::OperationCost;
+use storage::rocksdb_storage::RocksDbStorage;
+use crate::batch::key_info::KeyInfo;
+use crate::batch::KeyInfoPath;
+use crate::GroveDb;
+
+#[cfg(feature = "full")]
+impl GroveDb {
+    pub fn worst_case_for_has_raw(
+        path: &KeyInfoPath,
+        key: &KeyInfo,
+        max_element_size: u32,
+        in_parent_tree_using_sums: bool,
+    ) -> OperationCost {
+        let mut cost = OperationCost::default();
+        GroveDb::add_worst_case_has_raw_cost::<RocksDbStorage>(
+            &mut cost,
+            path,
+            key,
+            max_element_size,
+            in_parent_tree_using_sums,
+        );
+        cost
+    }
+
+    pub fn worst_case_for_get_raw(
+        path: &KeyInfoPath,
+        key: &KeyInfo,
+        max_element_size: u32,
+        in_parent_tree_using_sums: bool,
+    ) -> OperationCost {
+        let mut cost = OperationCost::default();
+        GroveDb::add_worst_case_get_raw_cost::<RocksDbStorage>(
+            &mut cost,
+            path,
+            key,
+            max_element_size,
+            in_parent_tree_using_sums,
+        );
+        cost
+    }
+
+    pub fn worst_case_for_get(
+        path: &KeyInfoPath,
+        key: &KeyInfo,
+        max_element_size: u32,
+        max_references_sizes: Vec<u32>,
+        in_parent_tree_using_sums: bool,
+    ) -> OperationCost {
+        let mut cost = OperationCost::default();
+        GroveDb::add_worst_case_get_cost::<RocksDbStorage>(
+            &mut cost,
+            path,
+            key,
+            max_element_size,
+            in_parent_tree_using_sums,
+            max_references_sizes,
+        );
+        cost
+    }
+}
